+++ base/build.gradle.kts
@@ -1,6 +1,5 @@
 plugins {
     kotlin("multiplatform")
-    kotlin("plugin.serialization")
     id("jacoco")
     id("convention.publication")
     id("com.android.library")
@@ -18,7 +17,6 @@ kotlin {
     sourceSets {
 
         val ext = (gradle as ExtensionAware).extra
-        val serialization_version: String by ext
 
         getByName("commonMain") {
             dependencies {
@@ -27,8 +25,6 @@ kotlin {
                 api("co.touchlab:stately-isolate:${ext["stately_isolate_version"]}")
                 api("co.touchlab:stately-iso-collections:${ext["stately_isolate_version"]}")
                 api("co.touchlab:stately-concurrency:${ext["stately_version"]}")
-                api("org.jetbrains.kotlinx:kotlinx-serialization-core:$serialization_version")
-                api("org.jetbrains.kotlinx:kotlinx-serialization-json:$serialization_version")
             }
         }
         getByName("commonTest") {
