+++ scientific/src/androidLibMain/AndroidManifest.xml
@@ -1,0 +1,2 @@
+<manifest package="com.splendo.kaluga.scientific"
+          xmlns:android="http://schemas.android.com/apk/res/android" />
