+++ settings.gradle.kts
@@ -70,4 +70,5 @@ include(":keyboard")
 include(":links")
 include(":resources")
 include(":review")
+include(":scientific")
 include(":system")
